rag/knowledge-bases/features-examples/04-infrastructure/e2e_rag_using_bedrock_kb_cdk/requirements.txt
aws-lambda-powertools
boto3
requests-aws4auth
opensearch-py
aws-cdk-lib